Extract Map component outside of the Main
